+++ java/fury-core/src/test/java/io/fury/resolver/ClassResolverTest.java
@@ -36,7 +36,7 @@ import io.fury.resolver.longlongpkg.C3;
 import io.fury.serializer.collection.CollectionSerializer;
 import io.fury.serializer.collection.CollectionSerializers;
 import io.fury.serializer.CompatibleSerializer;
-import io.fury.serializer.MapSerializers;
+import io.fury.serializer.map.MapSerializers;
 import io.fury.serializer.ObjectSerializer;
 import io.fury.serializer.Serializer;
 import io.fury.serializer.Serializers;
